% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRownameTable.R
\name{addRownameTable}
\alias{addRownameTable}
\title{addRownameTable}
\usage{
addRownameTable(tab, type = 1, samplename = "SampleName")
}
\arguments{
\item{tab}{The table to modify}

\item{type}{The type of modification: {0,3}=First column is indices, 1=First column is rownames, 2=1st column is indices AND 2nd column is rownames, 4=1st column is rownames AND collesponding column named "samplename"}

\item{samplename}{Stringname for "SampleName" (may be in another language)}
}
\value{
A modified table (wrt rownames)
}
\description{
Helpfunction for CaseSolver to set rownames as first column in table (modification)
}
