% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCONS.R
\name{calcCONS}
\alias{calcCONS}
\title{calcCONS}
\usage{
calcCONS(mlefitHp, mlefitHd, niter = 2000, delta = 2,
  quantile = 0.05, seed = NULL, verbose = TRUE,
  accRateGolden = 0.25, diffTol = 0.1)
}
\arguments{
\item{mlefitHp}{A fitted object returned from contLikMLE (under Hp)}

\item{mlefitHd}{A fitted object returned from contLikMLE (under Hd)}

\item{niter}{Number of samples in the MCMC-sampling.}

\item{delta}{A numerical parameter to scale with the covariance function Sigma. Default is 2. Should be higher to obtain lower acception rate.}

\item{quantile}{The quantile used to report conservative LR}

\item{seed}{The user can set seed if wanted}

\item{verbose}{Whether progress should be printed}

\item{accRateGolden}{aimed acceptance rate}

\item{diffTol}{Difference tolerance regarding acceptance rate}
}
\description{
A function for calculating conservative LR
}
\details{
The function calculates LR based on MCMC simulations
}
