% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEvidConc.R
\name{calcEvidConc}
\alias{calcEvidConc}
\title{calcEvidConc}
\usage{
calcEvidConc(nnTK, nLarge = 10000)
}
\arguments{
\item{nnTK}{an environment object from stored CaseSolver object (includes data)}

\item{nLarge}{A number indicating VERY large numeber of references}
}
\value{
A list with canidate results
}
\description{
Function for calculating evidence concordance
}
