% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHp.R
\name{calcHp}
\alias{calcHp}
\title{calcHp}
\usage{
calcHp(evidData, refData, nC, popFreq, kit, pC, lambda, threshT, xi,
  nDone = 2)
}
\arguments{
\item{evidData}{A list with evidence data}

\item{refData}{A list with reference data}

\item{nC}{Assumed number of contributors}

\item{popFreq}{A list with population frequences popFreq[[locusname]}

\item{kit}{Selected kit}

\item{pC}{Dropin probability}

\item{lambda}{Dropin PH hyperparameter}

\item{threshT}{Detection threshold used}

\item{xi}{Stutter proportion parameter}

\item{nDone}{Number of required optimisations}
}
\description{
A function to calculate under a specific hypotheses: a given evidence and given refData
}
