% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIBS.R
\name{calcIBS}
\alias{calcIBS}
\title{calcIBS}
\usage{
calcIBS(nnTK, nLarge = 10000, mixtureName = "mixture")
}
\arguments{
\item{nnTK}{an environment object from stored CaseSolver object (includes data)}

\item{nLarge}{A number indicating VERY large numeber of references}

\item{mixtureName}{A string indicating "mixture" (specific language)}
}
\value{
A list with canidate results
}
\description{
Calculate identical by state statistics between all references
}
