% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMACcomparison.R
\name{calcMACcomparison}
\alias{calcMACcomparison}
\title{calcMACcomparison}
\usage{
calcMACcomparison(DBmix, DBref, threshMAC)
}
\description{
A function calc number of alleles in ref-table which are within evidence-table
}
