% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcQuanLRcomparison.R
\name{calcQuanLRcomparison}
\alias{calcQuanLRcomparison}
\title{calcQuanLRcomparison}
\usage{
calcQuanLRcomparison(DBmix, DBref, matchlist, popFreq, kit, xi, pC, lambda,
  threshT, nDone = 4, maxC = 3, nContr = NULL)
}
\arguments{
\item{DBmix}{A list with evidence information [[sample]][[loc]] = list(adata,hdata)}

\item{DBref}{A (nR x nL) matrix with reference information. LIST not efficient}

\item{matchlist}{A matrix with overview of what references are matches each of samples (SampleName,Reference,MAC)}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{kit}{Used to model degradation. Must be one of the shortnames of kit: Returned by euroformix::getKit()}

\item{xi}{Model parameter for Backward stutter. Default is 0 (no stutter). NULL indiciates stutter model.}

\item{pC}{A numeric for allele drop-in probability. Default is 0.}

\item{lambda}{Parameter in modeled peak height shifted exponential model. Default is 0.}

\item{threshT}{The detection threshold given. Used when considering probability of allele drop-outs.}

\item{nDone}{Maximum number of random evaluations nlm-optimizing routing. Default is 4.}

\item{maxC}{Maximum number of contributors possible to assume. Default is 3.}

\item{nContr}{A vector of number of contributors for each of the row in the matchlist. If NULL (default), the max(nA)/2 rule is applied.}
}
\description{
Function calculating LR using EuroForMix for remaining matches (from Step 1)
}
