% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcQuanMLE.R
\name{calcQuanMLE}
\alias{calcQuanMLE}
\title{calcQuanMLE}
\usage{
calcQuanMLE(
  evidData,
  refData,
  condOrder,
  NOC,
  nnTK,
  isWOE = FALSE,
  verbose = FALSE,
  knownRef = NULL
)
}
\arguments{
\item{evidData}{A list with evidence data evidData[[sample]][[locus]]}

\item{refData}{A list with reference data refData[[refname]][[locname]]$adata}

\item{condOrder}{Conditional vector (hypothesis)}

\item{NOC}{Assumed number of contributors}

\item{nnTK}{an environment object from stored CaseSolver object}

\item{isWOE}{Indicating whether fst should be used or not}

\item{verbose}{Whether progress should be printed}

\item{knownRef}{Specify known non-contributing references from refData (index). For instance knownRef=(1,2) means that reference 1 and 2 is known non-contributor in the hypothesis. This affectes coancestry correction.}
}
\description{
A function to calculate under a specific hypotheses: a given evidence and given refData
}
