% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRMP.R
\name{calcRMP}
\alias{calcRMP}
\title{calcRMP}
\usage{
calcRMP(nnTK, sig = 3)
}
\arguments{
\item{nnTK}{an environment object from stored CaseSolver object}

\item{sig}{Number of significant values to use}
}
\value{
A list with elements evid and ref
}
\description{
Random match probability calculations
}
\details{
A function for calculating Random match probabilities/RMNE for each profiles (Evidence/Refs)
}
