% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWOEhyps.R
\name{calcWOEhyps}
\alias{calcWOEhyps}
\title{calcWOEhyps}
\usage{
calcWOEhyps(nnTK, useEmpty = FALSE, verbose = TRUE)
}
\arguments{
\item{nnTK}{an environment object from stored CaseSolver object}

\item{useEmpty}{Whether an empty hypotheses window should be shown (used to add more calculations)}

\item{verbose}{Whether progress should be printed}
}
\description{
A function for designing WoE hypothesis and calculate these, also statements are generated
}
\details{
The WOE window must be closed in order to return from function (wait flag)
}
