% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canPrintEPG.R
\name{canPrintEPG}
\alias{canPrintEPG}
\title{canPrintEPG}
\usage{
canPrintEPG(nnTK)
}
\arguments{
\item{nnTK}{an environment object from stored CaseSolver object}
}
\description{
A helpfunction to check if EPGs can be printed (kit specified)
}
