% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterOrder.R
\name{clusterOrder}
\alias{clusterOrder}
\title{clusterOrder}
\usage{
clusterOrder(table, plot = FALSE)
}
\arguments{
\item{table}{Table with genotypes}

\item{plot}{Whether the dendrogram should be plotted}
}
\value{
order vector (ranked wrt hierarchical clustering)
}
\description{
Order samples wrt to clustered hierachical structure (similarity)
}
\details{
Create distance between all genotypes (over all markers) and perform hierachical clustering (order returned)
DISTANCE OUTCOMES:
All shared (dist=0): (A/B vs A/B), (A/A vs A/A), (A vs A/B), (A vs A/A), (A vs A)
None shared (dist=2): (A/B vs C/D), (A/A vs B/B), (A vs B/C), (A vs B/B), (A vs B)
One shared (dist=1): (A/B vs B/C), (A/A vs A/B)
}
