% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createReport.R
\name{createReport}
\alias{createReport}
\title{createReport}
\usage{
createReport(nnTK)
}
\arguments{
\item{nnTK}{an environment object from stored CaseSolver object}
}
\description{
A function for creating report based on CaseSolver object
}
\details{
The function returns the names of the selected profiles (when window is closed)
}
