% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnvirKit.R
\name{getEnvirKit}
\alias{getEnvirKit}
\title{getEnvirKit}
\usage{
getEnvirKit(nnTK)
}
\arguments{
\item{nnTK}{an environment object from stored CaseSolver object}
}
\value{
Name of kit if found, otherwise NULL is returned.
}
\description{
Obtain defined kit from environment
}
