% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFreqs.R
\name{getFreqs}
\alias{getFreqs}
\title{getFreqs}
\usage{
getFreqs(filename)
}
\arguments{
\item{filename}{A string giving the file with allelefrequencies}
}
\description{
get allelefreq-info from file
}
