% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIBSdistr.R
\name{getIBSdistr}
\alias{getIBSdistr}
\title{getIBSdistr}
\usage{
getIBSdistr(popFreq, verbose = TRUE)
}
\arguments{
\item{popFreq}{A list with population frequences popFreq[[locusname]}

\item{verbose}{Boolean of whether progress should be printed to console}
}
\description{
A function to get the distribution of IBS for random individuals
}
