% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLanguage.R
\name{getLanguage}
\alias{getLanguage}
\title{getLanguage}
\usage{
getLanguage(langSel = NULL, encoding = "unknown",
  langFile = "Language", defaultLanguage = "English")
}
\arguments{
\item{langSel}{Selected language (NULL returns list of available languages)}

\item{encoding}{Encoding used to read text file (assuming english only)}

\item{langFile}{Name of language files (without extensions)}

\item{defaultLanguage}{Default language}
}
\description{
A function to extract list of words for selected language (files found in installation folder)
}
