% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelSettings.R
\name{getModelSettings}
\alias{getModelSettings}
\title{getModelSettings}
\usage{
getModelSettings(nnTK)
}
\arguments{
\item{nnTK}{an environment object from stored CaseSolver object}
}
\description{
Obtain model settings from environment
}
