% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleType2.R
\name{getSampleType2}
\alias{getSampleType2}
\title{getSampleType2}
\usage{
getSampleType2(samples, kit = NULL, LUSsymbol = "_")
}
\arguments{
\item{samples}{A list of samples (evidence) with structure [[samplename]][[locus]] = list(adata,hdata)}

\item{kit}{Kitname returned from getKit()}

\item{LUSsymbol}{Symbol to be recognized as a LUS allele}
}
\description{
A helpfunction for obtaining sample type
}
\details{
The function determines the type in the samples object ("EPG","LUS" or "MPS")
}
