% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStructuredData.R
\name{getStructuredData}
\alias{getStructuredData}
\title{getStructuredData}
\usage{
getStructuredData(data, ln, minLoc = 10, sortWithComplexity = FALSE)
}
\arguments{
\item{data}{a list with a evidence table (mix) and a reference table (ref)}

\item{ln}{markers to consider (a given order)}

\item{minLoc}{is minimum number of overlapping markers to be identical (for partial profiles).}

\item{sortWithComplexity}{Whether evidence profiles should be sorted with respect to complexity}
}
\description{
A function for structuring data
}
\details{
The function must return in a specific format.
}
