% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{gui}
\alias{gui}
\title{gui}
\usage{
gui(envirfile = NULL, envir = NULL)
}
\arguments{
\item{envirfile}{A file to a saved environment of a project (must contain nnTK)}

\item{envir}{A saved environment (nnTK)}
}
\description{
gui (GUI) is a GUI wrapper for CaseSolver
}
\details{
The function starts the CaseSolver software.
}
