% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number2word.R
\name{number2word}
\alias{number2word}
\title{number2word}
\usage{
number2word(x, L, prefix = "numword_", suffix = "numwordPlural")
}
\arguments{
\item{x}{LR number (on log10 scale)}

\item{L}{Language object (returned from getLanguage)}

\item{prefix}{Defined key-prefix for obtaining word equivalents}

\item{suffix}{Defined key for end of word when value>1 (plural)}
}
\value{
A vector with two elements (verbal and numeric expression)
}
\description{
Convert number to verbal
}
