% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderTableSort.R
\name{orderTableSort}
\alias{orderTableSort}
\title{orderTableSort}
\usage{
orderTableSort(var1, var2 = NULL, sort = 1)
}
\arguments{
\item{var1}{Variable to use for sorting (sort=2)}

\item{var2}{Additional variable to sort wrt (sort=3)}

\item{sort}{Type of sorting}
}
\description{
helpfunction to obtain order of a table
}
