% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileSelectorGUI.R
\name{profileSelectorGUI}
\alias{profileSelectorGUI}
\title{profileSelectorGUI}
\usage{
profileSelectorGUI(env = NULL)
}
\arguments{
\item{env}{an environment with elements evids and refs}
}
\description{
A function for letting user selecting profiles
}
\details{
The function returns the names of the selected profiles (when window is closed)
}
\examples{
\dontrun{
env = new.env( parent = emptyenv() ) 
assign("selected",list(evids=1:10,refs=1:100),envir=env)
profileSelectorGUI(env)
}
}
