% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileSwapperGUI.R
\name{profileSwapperGUI}
\alias{profileSwapperGUI}
\title{profileSwapperGUI}
\usage{
profileSwapperGUI(env)
}
\arguments{
\item{env}{an environment with elements items1 and items2}
}
\description{
User select items between two tables
}
\details{
The function returns the items selected in the two lists (when window is closed)
}
\examples{
\dontrun{
env = new.env( parent = emptyenv() ) 
assign("itemList",list(items1=paste("item",1:10),items2=paste("item",21:100)),envir=env)
profileSwapperGUI(env)
}
}
