% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_listToTable.R
\name{sample_listToTable}
\alias{sample_listToTable}
\title{sample_listToTable}
\usage{
sample_listToTable(Y, PH = TRUE)
}
\description{
A function for converting profiles from list to table.
}
