% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleListToTable.R
\name{sampleListToTable}
\alias{sampleListToTable}
\title{sampleListToTable}
\usage{
sampleListToTable(outL, PH = TRUE)
}
\arguments{
\item{outL}{A list on form outL[[samplename]][[locusname]]$adata,outL[[samplename]][[locusname]]$hdata (Evid or refs)}

\item{PH}{Boolean of whether Peak heights should be exported (used for LRmix studio)}
}
\value{
table
}
\description{
A function for converting profiles from list to table.

Converting profiles from list to table format (helpfunction)
}
