% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMarkerSettings.R
\name{setMarkerSettings}
\alias{setMarkerSettings}
\title{setMarkerSettings}
\usage{
setMarkerSettings(nnTK)
}
\arguments{
\item{nnTK}{an environment object from stored CaseSolver object}
}
\description{
Defining marker settings
}
\details{
The user defines marker settings by inserting values to table
}
