% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMatchNetwork.R
\name{showMatchNetwork}
\alias{showMatchNetwork}
\title{showMatchNetwork}
\usage{
showMatchNetwork(nnTK, action, createInteractive = FALSE,
  selList = NULL)
}
\arguments{
\item{nnTK}{an environment object from stored CaseSolver object (includes data)}

\item{action}{The type of plot to be produced (all,onlymix,onlyss)}

\item{createInteractive}{A boolean of whether to create an interactive matchnetwork (shown in browser)}

\item{selList}{A list (evids,refs) to only show the matchnetwork for. Default is not considered}
}
\value{
A list with canidate results
}
\description{
Calculate identical by state statistics between all references
}
