% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHp.R
\name{calcHp}
\alias{calcHp}
\title{calcHp}
\usage{
calcHp(evidData, refData, nC, popFreq, kit, pC, lambda, threshT, nDone = 4)
}
\description{
A function to calculate under a specific hypotheses: a given evidence and given refData
}
