% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMACcomparison.R
\name{calcMACcomparison}
\alias{calcMACcomparison}
\title{calcMACcomparison}
\usage{
calcMACcomparison(DBmix, DBref, threshMAC)
}
\arguments{
\item{DBmix}{A (nS x nL) matrix with evidence (only alleles) information. LIST not efficient}

\item{DBref}{A (nR x nL) matrix with reference information. LIST not efficient}

\item{threshMAC}{Required MAC threshold for being a candidate match}
}
\description{
A function calc number of alleles in ref-table which are within evidence-table
}
