% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcQualLRcomparison.R
\name{calcQualLRcomparison}
\alias{calcQualLRcomparison}
\title{calcQualLRcomparison}
\usage{
calcQualLRcomparison(DBmix, DBref, matchlist, popFreq, pC = 0.05,
  pD0 = 0.1, maxC = 6)
}
\arguments{
\item{DBmix}{A list with evidence information [[sample]][[loc]] = list(adata,hdata)}

\item{DBref}{A (nR x nL) matrix with reference information. LIST not efficient}

\item{matchlist}{A matrix with overview of what references are matches each of samples (SampleName,Reference)}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{pC}{A numeric for allele drop-in probability. Default is 0.}

\item{pD0}{Initiative dropout parameter}

\item{maxC}{Maximum number of contributors possible to assume. Default is 6.}
}
\description{
Function calculating LR using EuroForMix for remaining matches (from Step 1)
}
