% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcQuanLRcomparison.R
\name{calcQuanLRcomparison}
\alias{calcQuanLRcomparison}
\title{calcQuanLRcomparison}
\usage{
calcQuanLRcomparison(
  DBmix,
  DBref,
  matchlist,
  popFreq,
  kit,
  xiBW = 0,
  xiFW = 0,
  pC = 0.05,
  lambda = 0.01,
  threshT = 50,
  nDone = 4,
  maxC = 3,
  nContr = NULL,
  normalize = FALSE,
  minFreq = NULL,
  useEFMex = FALSE
)
}
\arguments{
\item{DBmix}{A list with evidence information [[sample]][[loc]] = list(adata,hdata)}

\item{DBref}{A (nR x nL) matrix with reference information. LIST not efficient}

\item{matchlist}{A matrix with overview of what references are matches each of samples (SampleName,Reference,MAC)}

\item{popFreq}{A list of allele frequencies for a given population.}

\item{kit}{Used to model degradation. Must be one of the shortnames of kit: Returned by euroformix::getKit()}

\item{xiBW}{Model parameter for Backward stutter. NULL indicates stutter model.}

\item{xiFW}{Model parameter for Forward stutter. NULL indicates stutter model.}

\item{pC}{A numeric for allele drop-in probability.}

\item{lambda}{Parameter in modeled peak height shifted exponential model.}

\item{threshT}{The detection threshold given.}

\item{nDone}{Maximum number of random evaluations nlm-optimizing routing. Default is 4.}

\item{maxC}{Maximum number of contributors possible to assume. Default is 3.}

\item{nContr}{A vector of number of contributors for each of the row in the matchlist. If NULL (default), the max(nA)/2 rule is applied.}

\item{normalize}{A boolean of whether normalization should be applied or not.}

\item{minFreq}{The freq value included for new alleles. Default NULL is using min.observed in popFreq.}

\item{useEFMex}{Whether the exhaustive method should be used}
}
\description{
Function calculating LR using EuroForMix for remaining matches (from Step 1)
}
