% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabToListRef.R
\name{tabToListRef}
\alias{tabToListRef}
\title{tabToListRef}
\usage{
tabToListRef(tab, ln = NULL, setEmpty = FALSE)
}
\arguments{
\item{tab}{A (nR x nL) matrix with rows giving locus specific genotypes (for each column) for references}

\item{ln}{Loci names of return list}

\item{setEmpty}{Whether genotypes with one allele is set as empty or not (then set as homozygous)}
}
\value{
List format of references refL[[refname]][[locname]]$adata
}
\description{
Function to convert a matrix to list format.
}
\details{
Used to convert reference data stored in matrix to list format required by functions.
}
